mixed formatting for tpane
