+++ src/js/CheckModule.js
@@ -5,7 +5,6 @@ var CheckModule = React.createClass({
   render: function() {
     return (
       <div>
-        {/* rendering the buttons and passing the call backs as props */ }
         <RetainedButton onCheckedStatusChanged={this.props.onCheckedStatusChanged} />
         <ReplacedButton onCheckedStatusChanged={this.props.onCheckedStatusChanged} />
         <WrongButton onCheckedStatusChanged={this.props.onCheckedStatusChanged} />
