+++ src/js/Checks/tpane.js
@@ -10,19 +10,12 @@ const Well = require('react-bootstrap/lib/Well.js');
 const React = require('react');
 const FileActions = require('../FileActions');
 
-var style = {
-  content: {
-    paddingLeft: 3,
-    paddingRight: 3
-  }
-};
-
 var Pane = React.createClass({
   render: function() {
     return (
       <div>
           <h3>{this.props.title}</h3>
-         <Well>{this.props.content}</Well>
+         <Well className="tpane" >{this.props.content}</Well>
       </div>
     );
   }
@@ -47,10 +40,10 @@ var TPane = React.createClass({
   render: function() {
     return (
       <Grid>
-      <Row style={style.content}>
-          <Col style={style.content} xs={3} md={3}><Pane title="Original Language" content={this.state.ol}/></Col>
-          <Col style={style.content} xs={3} md={3}><Pane title="Gateway Language" content={this.state.gl}/></Col>
-          <Col style={style.content} xs={3} md={3}><Pane title="Target Language" content={this.state.tl}/></Col>
+      <Row >
+          <Col xs={3} md={3}><Pane title="Original Language" content={this.state.ol}/></Col>
+          <Col xs={3} md={3}><Pane title="Gateway Language" content={this.state.gl}/></Col>
+          <Col xs={3} md={3}><Pane title="Target Language" content={this.state.tl}/></Col>
       </Row>
       </Grid>
   );
